+++ pdf_editor_offline/utils/validators.py
@@ -1,6 +1,6 @@
 import os
 
-import fitz
+import pymupdf as fitz
 from PIL import Image